+++ snowflake/ml/modeling/impute/simple_imputer.py
@@ -49,6 +49,7 @@ _SKLEARN_UNUSED_KEYWORDS = [
 _SNOWML_ONLY_KEYWORDS = [
     "input_cols",
     "output_cols",
+    "passthrough_cols",
 ]  # snowml only keywords not present in sklearn
 
 # Added keywords mapped to the sklearn versions in which they were added. Update mappings in new
@@ -106,6 +107,11 @@ class SimpleImputer(base.BaseTransformer):
         output_cols: Optional[Union[str, List[str]]]
             A string or list of strings representing column names that will store the output of transform operation.
             The length of `output_cols` must equal the length of `input_cols`.
+        passthrough_cols: A string or a list of strings indicating column names to be excluded from any
+            operations (such as train, transform, or inference). These specified column(s)
+            will remain untouched throughout the process. This option is helpful in scenarios
+            requiring automatic input_cols inference, but need to avoid using specific
+            columns, like index columns, during training or inference.
         drop_input_cols: bool, default=False
             Remove input columns from output if set `True`.
 
@@ -130,6 +136,7 @@ class SimpleImputer(base.BaseTransformer):
         fill_value: Optional[Union[str, float]] = None,
         input_cols: Optional[Union[str, Iterable[str]]] = None,
         output_cols: Optional[Union[str, Iterable[str]]] = None,
+        passthrough_cols: Optional[Union[str, Iterable[str]]] = None,
         drop_input_cols: Optional[bool] = False,
     ) -> None:
         super().__init__(drop_input_cols=drop_input_cols)
@@ -158,6 +165,7 @@ class SimpleImputer(base.BaseTransformer):
 
         self.set_input_cols(input_cols)
         self.set_output_cols(output_cols)
+        self.set_passthrough_cols(passthrough_cols)
 
     def _reset(self) -> None:
         """