+++ snowflake/ml/modeling/preprocessing/robust_scaler.py
@@ -32,6 +32,11 @@ class RobustScaler(base.BaseTransformer):
         output_cols: The name(s) of one or more columns in a DataFrame in which results will be stored. The number of
             columns specified must match the number of input columns. For dense output, the column names specified are
             used as base names for the columns created for each category.
+        passthrough_cols: A string or a list of strings indicating column names to be excluded from any
+            operations (such as train, transform, or inference). These specified column(s)
+            will remain untouched throughout the process. This option is helpful in scenarios
+            requiring automatic input_cols inference, but need to avoid using specific
+            columns, like index columns, during training or inference.
         drop_input_cols: Remove input columns from output if set True. False by default.
 
     Attributes:
@@ -48,6 +53,7 @@ class RobustScaler(base.BaseTransformer):
         unit_variance: bool = False,
         input_cols: Optional[Union[str, Iterable[str]]] = None,
         output_cols: Optional[Union[str, Iterable[str]]] = None,
+        passthrough_cols: Optional[Union[str, Iterable[str]]] = None,
         drop_input_cols: Optional[bool] = False,
     ) -> None:
         """
@@ -68,6 +74,11 @@ class RobustScaler(base.BaseTransformer):
                 If less than 1, the dataset will be scaled up.
             input_cols: Single or multiple input columns.
             output_cols: Single or multiple output columns.
+            passthrough_cols: A string or a list of strings indicating column names to be excluded from any
+                operations (such as train, transform, or inference). These specified column(s)
+                will remain untouched throughout the process. This option is helful in scenarios
+                requiring automatic input_cols inference, but need to avoid using specific
+                columns, like index columns, during in training or inference.
             drop_input_cols: Remove input columns from output if set True. False by default.
 
         Attributes:
@@ -95,6 +106,7 @@ class RobustScaler(base.BaseTransformer):
 
         self.set_input_cols(input_cols)
         self.set_output_cols(output_cols)
+        self.set_passthrough_cols(passthrough_cols)
 
     def _reset(self) -> None:
         """